pyproject.toml
[project]
name = "excelforge"
version = "2.0.0"
description = "ExcelForge V2 multi-MCP runtime with shared JSON-RPC pipe backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8,<3",
  "PyYAML>=6.0.1,<7",
  "mcp>=1.2.0",
  "pywin32>=306; platform_system == 'Windows'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
  "pytest-cov>=5.0,<6",
]

[project.scripts]
excelforge = "excelforge.main:main"
excelforge-runtime = "excelforge.runtime.main:main"
excelforge-core = "excelforge.gateway.core.main:main"
excelforge-vba = "excelforge.gateway.vba.main:main"
excelforge-recovery = "excelforge.gateway.recovery.main:main"
excelforge-pq = "excelforge.gateway.pq.main:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["excelforge*"]

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
addopts = "-q"
markers = [
  "integration: requires local Excel Desktop with pywin32",
  "e2e: end-to-end MCP stdio tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
